+++ CHANGES.md
@@ -6,8 +6,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 Use this section to keep track of changes in the works.
+### Added
+* `rotate` method in SiteBasis class.
+
+### Changed
+
+### Fixed
+* Fix orbit generation to play niceley with changes in pymatgen
+  `Structure.sites_in_sphere` return value.
+  [\#125](https://github.com/CederGroupHub/smol/pull/125)
+  ([lbluque](https://github.com/lbluque))
+
+## [v1.0.1](https://github.com/CederGroupHub/smol/tree/v1.0.1) (2021-03-03)
+#### [Full Changelog](https://github.com/CederGroupHub/smol/compare/v1.0.0...v1.0.1)
 
 ### Added
+* Method in `StructureWrangler` to get structure matching duplicates
+  [\#122](https://github.com/CederGroupHub/smol/pull/122)
+  ([lbluque](https://github.com/lbluque))
+* Include tolerance when detecting duplicate correlation vectors.
+  [\#121](https://github.com/CederGroupHub/smol/pull/122)
+  ([lbluque](https://github.com/lbluque))
+* Convenience method to get feature matrix orbit ranks.
+  [\#117](https://github.com/CederGroupHub/smol/pull/117)
+  ([lbluque](https://github.com/lbluque))
 * bit combo hierarchy in `ClusterSubspace` for fitting hierarchy constraints.
   [\#106](https://github.com/CederGroupHub/smol/pull/106)
   ([qchempku2017](https://github.com/qchempku2017))
@@ -30,6 +52,12 @@ Use this section to keep track of changes in the works.
   - `orbits_by_cutoffs`, `function_inds_by_cutoffs`, `function_inds_by_size`.
 
 ### Changed
+* Allow using external term values when detecting duplicate corr vectors.
+[\#124](https://github.com/CederGroupHub/smol/pull/124)
+([lbluque](https://github.com/lbluque))
+* Warn instead of printing when structure matching fails.
+[\#124](https://github.com/CederGroupHub/smol/pull/124)
+([lbluque](https://github.com/lbluque))
 * filter functions in `smol.wrangling` replaced with functions returning
   indices corresponding to structures to keep. This can be used saving indices
   with `StructureWrangler.add_data_indices`.
@@ -55,6 +83,16 @@ Use this section to keep track of changes in the works.
 ([lbluque](https://github.com/lbluque))
  
 ### Fixed
+* Raise error in `StructureWrangler.append_data_items` when item properties are
+  missing keys already included.
+[\#117](https://github.com/CederGroupHub/smol/pull/117)
+  ([lbluque](https://github.com/lbluque))
+* Correctly recreate coefs in `CompositeProcessor.from_dict`
+[\#116](https://github.com/CederGroupHub/smol/pull/116)
+  ([lbluque](https://github.com/lbluque))
+* Disallow setting chemical potentials/fugacities with duplicate string/species
+  in dictionary. [\#114](https://github.com/CederGroupHub/smol/pull/114)
+  ([lbluque](https://github.com/lbluque))
 * Fixed loading `ClusterSubspace` with polynomial basis from dict.
 [\#112](https://github.com/CederGroupHub/smol/pull/112)
   ([lbluque](https://github.com/lbluque))
@@ -65,7 +103,7 @@ Use this section to keep track of changes in the works.
 sets. [\#90](https://github.com/CederGroupHub/smol/pull/90)
   ([lbluque](https://github.com/lbluque))
 
-## [v1.0.0]() (2020-10-27)
+## [v1.0.0](https://github.com/CederGroupHub/smol/releases/tag/v1.0.0) (2020-10-27)
 #### [Full Changelog](https://github.com/CederGroupHub/smol/compare/v0.0.0...v1.0.0)
 ### Added
 * Completely new `smol.moca` module. Design based generally up as follows: